task 6

+++ 0x00-hello_world/6-size.c
@@ -5,11 +5,11 @@
  */
 int main(void)
 {
-print f("Size of a char: %lu byte(s)\n", (unsigned long)sizeof(char));
-print("Size of an int: %lu byte(s)\n", (unsigned long)sizeof(int));
-print("Size of a long int: %lu byte(s)\n", (unsigned long)sizeof(long int));
-print("Size of a long long int: %lu byte(s)\n", (unsigned long)sizeof(long long int));
-printf("Size of a float: %lu byte(s)\n", (unsigned long)sizeof(float));
+print f("Size of a char: %lu byte(s)\n", sizeof(char));
+print("Size of an int: %lu byte(s)\n", sizeof(int));
+print("Size of a long int: %lu byte(s)\n", sizeof(long int));
+print("Size of a long long int: %lu byte(s)\n", sizeof(long long int));
+printf("Size of a float: %lu byte(s)\n", sizeof(float));
 return (0);
 }
